+++ packages/story-editor/src/components/panels/document/taxonmies/stories/index.js
@@ -13,5 +13,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/**
+ * External dependencies
+ */
+import styled from 'styled-components';
+/**
+ * Internal dependencies
+ */
+import TaxonomiesPanel from '../taxonomies';
 
-export { default } from './categories';
+const Wrapper = styled.div`
+  background: ${({ theme }) => theme.colors.bg.primary};
+`;
+
+export default {
+  title: 'Stories Editor/Components/Taxonomies',
+  component: TaxonomiesPanel,
+};
+
+function Taxonomies() {
+  return <TaxonomiesPanel />;
+}
+
+export const _default = () => {
+  return (
+    <Wrapper>
+      <Taxonomies />
+    </Wrapper>
+  );
+};
